Grid check added
Added a check to insure extended 6 digit gridsquare updates from qrz or hamqth are consistent with 4 digit logged gridsquares. If they are not it may mean the contact was operating from a remote location, and so the gridsquare should not be updated in that case. Mis-matches are logged in the error file.